android/requirements.txt
behave
browserstack-local
paver
selenium
psutil
Appium-Python-Client
